refactor: eliminate marketing optimization quality regressions
