+++ .agents/scripts/marketing_experiment_assignment.py
@@ -10,6 +10,14 @@ from dataclasses import dataclass
 from typing import Any
 
 from marketing_experiment_evidence import AssignmentEvidence, load_assignment_evidence
+from marketing_experiment_assignment_grouping import (
+    GroupingContext,
+    contaminated_subjects,
+    group_post_exposure_events,
+    observed_labels,
+    post_exposure_assignments,
+    subject_variant,
+)
 from marketing_optimization_contract import (
     OptimizationError,
     OptimizationSnapshot,
@@ -78,143 +86,11 @@ class ExperimentPopulation:
     assignment_verified: bool
 
 
-@dataclass(frozen=True)
-class GroupedEvents:
-    """Post-exposure aggregate inputs and exclusion signals."""
-
-    by_variant: dict[str, list[dict[str, Any]]]
-    exposed: dict[str, set[str]]
-    pre_exposure_excluded: bool
-    unexposed_excluded: bool
-
-
-@dataclass(frozen=True)
-class GroupingContext:
-    """Assignment and exclusion state used while grouping events."""
-
-    variant_ids: set[str]
-    excluded: set[str]
-    labels: dict[str, set[str]]
-    evidence: AssignmentEvidence
-    exposure_metric: str
-
-
 def _within_window(event: dict[str, Any], started_at: str, ended_at: str, as_of: str) -> bool:
     """Return whether one event is observable inside the experiment window."""
     occurred = parse_datetime(event["event"]["occurred_at"], "experiment event occurred_at")
     upper = min(parse_datetime(ended_at, "experiment ended_at"), parse_datetime(as_of, "analysis as_of"))
     return parse_datetime(started_at, "experiment started_at") <= occurred <= upper
-
-
-def _observed_labels(
-    events: tuple[dict[str, Any], ...],
-    variant_ids: set[str],
-) -> dict[str, set[str]]:
-    """Collect valid observed variant labels by pseudonymous subject."""
-    labels: dict[str, set[str]] = {}
-    for event in events:
-        subject_id = event["subject"].get("subject_id")
-        variant = event["scope"].get("dimensions", {}).get("experiment_variant")
-        if subject_id is None or variant not in variant_ids:
-            continue
-        labels.setdefault(str(subject_id), set()).add(str(variant))
-    return labels
-
-
-def _contaminated_subjects(
-    labels: dict[str, set[str]],
-    evidence: AssignmentEvidence,
-) -> set[str]:
-    """Detect crossovers and labels that contradict verified assignment."""
-    contaminated = {subject_id for subject_id, variants in labels.items() if len(variants) > 1}
-    if evidence.verified:
-        contaminated.update(
-            subject_id
-            for subject_id, variants in labels.items()
-            if subject_id in evidence.assignments and variants != {evidence.assignments[subject_id]}
-        )
-    return contaminated
-
-
-def _post_exposure_assignments(
-    events: tuple[dict[str, Any], ...],
-    evidence: AssignmentEvidence,
-    exposure_metric: str,
-) -> set[str]:
-    """Reject exposures that predate their verified sticky assignment."""
-    if not evidence.verified:
-        return set()
-    contaminated: set[str] = set()
-    for event in events:
-        subject_id = event["subject"].get("subject_id")
-        if subject_id not in evidence.assigned_at or event["measurement"]["metric_id"] != exposure_metric:
-            continue
-        occurred = parse_datetime(event["event"]["occurred_at"], "exposure occurred_at")
-        assigned = parse_datetime(evidence.assigned_at[str(subject_id)], "assignment assigned_at")
-        if occurred < assigned:
-            contaminated.add(str(subject_id))
-    return contaminated
-
-
-def _subject_variant(
-    subject_id: str,
-    labels: dict[str, set[str]],
-    evidence: AssignmentEvidence,
-) -> str | None:
-    """Use verified assignment or one unambiguous observational label."""
-    if evidence.verified:
-        return evidence.assignments.get(subject_id)
-    observed = labels.get(subject_id, set())
-    return next(iter(observed)) if len(observed) == 1 else None
-
-
-def _first_exposures(
-    events: tuple[dict[str, Any], ...],
-    exposure_metric: str,
-) -> dict[str, Any]:
-    """Return the first observed exposure time for each pseudonymous subject."""
-    first: dict[str, Any] = {}
-    for event in events:
-        subject_id = event["subject"].get("subject_id")
-        if subject_id is None or event["measurement"]["metric_id"] != exposure_metric:
-            continue
-        occurred = parse_datetime(event["event"]["occurred_at"], "exposure occurred_at")
-        current = first.get(str(subject_id))
-        if current is None or occurred < current:
-            first[str(subject_id)] = occurred
-    return first
-
-
-def _group_post_exposure_events(
-    events: tuple[dict[str, Any], ...],
-    context: GroupingContext,
-) -> GroupedEvents:
-    """Group only events observed on or after each subject's first exposure."""
-    first_exposures = _first_exposures(events, context.exposure_metric)
-    by_variant: dict[str, list[dict[str, Any]]] = {variant_id: [] for variant_id in context.variant_ids}
-    exposed: dict[str, set[str]] = {variant_id: set() for variant_id in context.variant_ids}
-    pre_exposure_excluded = False
-    unexposed_excluded = False
-    for event in events:
-        subject_id = event["subject"].get("subject_id")
-        if subject_id is None or subject_id in context.excluded:
-            continue
-        subject_ref = str(subject_id)
-        variant = _subject_variant(subject_ref, context.labels, context.evidence)
-        if variant is None:
-            continue
-        first_exposure = first_exposures.get(subject_ref)
-        if first_exposure is None:
-            unexposed_excluded = True
-            continue
-        occurred = parse_datetime(event["event"]["occurred_at"], "experiment event occurred_at")
-        if occurred < first_exposure:
-            pre_exposure_excluded = True
-            continue
-        by_variant[variant].append(event)
-        if event["measurement"]["metric_id"] == context.exposure_metric:
-            exposed[variant].add(subject_ref)
-    return GroupedEvents(by_variant, exposed, pre_exposure_excluded, unexposed_excluded)
 
 
 def build_population(
@@ -245,14 +121,14 @@ def build_population(
         for event in window_events
         if not identity_is_uncertain(event["subject"].get("identity_state"))
     )
-    labels = _observed_labels(events, variant_ids)
+    labels = observed_labels(events, variant_ids)
     exposure_metric = str(assignment.get("exposure_metric_id"))
-    contaminated = _contaminated_subjects(labels, canonical_evidence)
-    contaminated.update(_post_exposure_assignments(events, canonical_evidence, exposure_metric))
+    contaminated = contaminated_subjects(labels, canonical_evidence)
+    contaminated.update(post_exposure_assignments(events, canonical_evidence, exposure_metric))
     invalid = bool(contaminated) and assignment.get("contamination_policy") == "mark_invalid"
     excluded = set(contaminated).union(uncertain_subjects)
     context = GroupingContext(variant_ids, excluded, labels, canonical_evidence, exposure_metric)
-    grouped = _group_post_exposure_events(events, context)
+    grouped = group_post_exposure_events(events, context)
     reasons = set(canonical_evidence.reasons)
     if contaminated:
         reasons.add("assignment_contamination" if invalid else "crossovers_excluded")
@@ -262,7 +138,7 @@ def build_population(
         str(event["subject"]["subject_id"])
         for event in events
         if event["subject"].get("subject_id") is not None
-        and _subject_variant(str(event["subject"]["subject_id"]), labels, canonical_evidence) is None
+        and subject_variant(str(event["subject"]["subject_id"]), labels, canonical_evidence) is None
     }
     if unassigned:
         reasons.add("unassigned_subjects_excluded")
