+++ .agents/scripts/tests/test-marketing-optimization.py
@@ -5,7 +5,6 @@
 
 from __future__ import annotations
 
-import hashlib
 import json
 import os
 import stat
@@ -21,7 +20,7 @@ from unittest import mock
 SCRIPTS = Path(__file__).resolve().parents[1]
 sys.path.insert(0, str(SCRIPTS))
 
-from marketing_attribution import AttributionRequest, _last_touch_key, build_attribution  # noqa: E402
+from marketing_attribution import _last_touch_key, build_attribution  # noqa: E402
 from marketing_attribution_validation import validate_attribution_artifact  # noqa: E402
 from growth_recommendations import (  # noqa: E402
     RecommendationPolicy,
@@ -43,7 +42,26 @@ from marketing_optimization_artifact_registry import (  # noqa: E402
     registered_recommendation,
     registered_report,
 )
-from marketing_optimization_test_support import aggregate_snapshot_document  # noqa: E402
+from marketing_optimization_test_support import (  # noqa: E402
+    ACCOUNT,
+    AS_OF,
+    CAMPAIGN,
+    aggregate_analysis_document,
+    aggregate_snapshot_document,
+    analysis_document,
+    attribution_request,
+    experiment_event,
+    experiment_fixture,
+    future_cli_experiment_fixture,
+    hashed_reference,
+    load_fixture,
+    mark_uncertain_identities,
+    normalized_event,
+    reference,
+    snapshot_document,
+    subject,
+)
+from marketing_optimization_snapshot_test_mixin import SnapshotAttributionMixin  # noqa: E402
 from marketing_optimization_contract import (  # noqa: E402
     MINIMUM_AGGREGATE_CELL_SIZE,
     MINIMUM_EXPERIMENT_CONVERSIONS_PER_VARIANT,
@@ -91,341 +109,13 @@ from performance_reporting import PerformanceReporting  # noqa: E402
 from performance_store import MarketingPerformanceStore  # noqa: E402
 from marketing_report_validation import validate_report_artifact  # noqa: E402
 
-AS_OF = "2026-08-15T00:00:00Z"
-ACCOUNT = "fixture-account"
-CAMPAIGN = "growth-campaign"
 ATTRIBUTION_SCHEMA = SCRIPTS.parent / "schemas" / "marketing-attribution.schema.json"
 EXPERIMENT_SCHEMA = SCRIPTS.parent / "schemas" / "marketing-experiment.schema.json"
 RECOMMENDATION_SCHEMA = SCRIPTS.parent / "schemas" / "growth-recommendation.schema.json"
 HELPER = SCRIPTS / "marketing-optimization-helper.py"
-FIXTURES = Path(__file__).resolve().parent / "fixtures" / "marketing-optimization"
 
 
-def load_fixture(name: str) -> dict[str, object]:
-    """Load one committed, synthetic optimization fixture."""
-    return json.loads((FIXTURES / name).read_text(encoding="utf-8"))
-
-
-def reference(prefix: str, digit: str) -> str:
-    """Return one stable synthetic typed reference."""
-    return f"{prefix}:{digit * 64}"
-
-
-def normalized_event(
-    digit: str,
-    event_type: str,
-    occurred_at: str,
-    *,
-    subject_id: str | None = None,
-) -> dict[str, object]:
-    """Build one schema-shaped normalized event without provider data."""
-    metric_id, unit = {
-        "engagement": ("marketing.engagement.total", "engagement"),
-        "impression": ("marketing.impressions.total", "impression"),
-        "visit": ("marketing.visits.total", "visit"),
-    }.get(event_type, ("marketing.conversions.total", "conversion"))
-    return {
-        "schema_version": 1,
-        "record_ref": reference("mkt-record-v1", digit),
-        "event_ref": reference("mkt-event-v1", digit),
-        "source": {
-            "kind": "normalized",
-            "account_ref": ACCOUNT,
-            "revision": 1,
-            "observed_at": occurred_at,
-            "recorded_at": occurred_at,
-            "source_observed_at": occurred_at,
-            "source_recorded_at": occurred_at,
-            "evidence_ref": reference("mkt-evidence-v1:sha256", digit),
-            "coverage": "complete",
-            "missing_scopes": [],
-        },
-        "event": {"type": event_type, "occurred_at": occurred_at, "correction_of": None},
-        "subject": {
-            "subject_id": subject_id,
-            "kind": "lead" if subject_id else "aggregate",
-            "identity_state": "isolated" if subject_id else "not_applicable",
-        },
-        "scope": {
-            "campaign_id": CAMPAIGN,
-            "channel": "search",
-            "creative_id": "creative-a",
-            "touchpoint_id": "touch-a" if event_type in {"engagement", "visit"} else None,
-            "outcome_id": "outcome-a" if event_type not in {"engagement", "visit"} else None,
-            "dimensions": {},
-        },
-        "measurement": {
-            "metric_id": metric_id,
-            "value": 1,
-            "unit": unit,
-            "aggregation": "sum",
-            "currency": None,
-        },
-        "quality": {
-            "confidence": "high",
-            "effective_confidence": "high",
-            "completeness": "complete",
-            "source_type": "fixture",
-            "collected_by": "synthetic-fixture",
-            "evidence_ref": reference("mkt-evidence-v1:sha256", digit),
-        },
-        "governance": {"audience_eligible": False, "eligibility_reason": "consent_unknown"},
-    }
-
-
-def subject(digit: str) -> dict[str, object]:
-    """Build one pseudonymous subject projection."""
-    subject_id = reference("mkt-subj-v1", digit)
-    return {
-        "schema_version": 1,
-        "subject_id": subject_id,
-        "kind": "lead",
-        "identity_state": "isolated",
-        "canonical_subject_id": subject_id,
-        "aliases": [subject_id],
-        "consent": [],
-        "suppression": [],
-        "identity_history": [],
-        "audience_eligible": False,
-        "eligibility_reason": "consent_unknown",
-    }
-
-
-def snapshot_document() -> dict[str, object]:
-    """Return one deterministic fixture with an out-of-bound future event."""
-    return load_fixture("sparse-journey.json")
-
-
-def analysis_document() -> dict[str, object]:
-    """Return a mature, fresh snapshot for deterministic analysis tests."""
-    document = snapshot_document()
-    document["events"] = document["events"][:2]
-    document["sources"][0]["last_observed_at"] = AS_OF
-    document["sources"][0]["stale"] = False
-    document["sources"][0]["lag_seconds"] = 0
-    return document
-
-
-def mark_uncertain_identities(document: dict[str, object], identity_state: str) -> dict[str, object]:
-    """Mark every synthetic subject as non-distinct identity evidence."""
-    for event in document["events"]:
-        event["subject"]["identity_state"] = identity_state
-        event["governance"] = {
-            "audience_eligible": False,
-            "eligibility_reason": "identity_ambiguous",
-        }
-    for item in document["subjects"]:
-        item["identity_state"] = identity_state
-        item["audience_eligible"] = False
-        item["eligibility_reason"] = "identity_ambiguous"
-    return document
-
-
-def aggregate_analysis_document() -> dict[str, object]:
-    """Return a visible aggregate while preserving the sparse source fixture."""
-    return aggregate_snapshot_document(analysis_document())
-
-
-def attribution_request(**overrides: object) -> AttributionRequest:
-    """Return one compact request at the immutable privacy floor."""
-    values: dict[str, object] = {
-        "outcome_metric_id": "marketing.conversions.total",
-        "model": "last_touch",
-        "lookback_seconds": 30 * 86400,
-        "refund_maturity_seconds": 0,
-        "minimum_cell_size": MINIMUM_AGGREGATE_CELL_SIZE,
-    }
-    values.update(overrides)
-    return AttributionRequest(**values)
-
-
-def hashed_reference(prefix: str, label: str) -> str:
-    """Return a valid deterministic typed reference for larger fixtures."""
-    return f"{prefix}:{hashlib.sha256(label.encode('utf-8')).hexdigest()}"
-
-
-def experiment_event(index: int, spec: dict[str, object]) -> dict[str, object]:
-    """Build one subject-level experiment event from compact fixture fields."""
-    event_type = str(spec["event_type"])
-    metric_id = str(spec["metric_id"])
-    subject_id = str(spec["subject_id"])
-    variant_id = str(spec["variant_id"])
-    event = normalized_event("1", event_type, str(spec.get("occurred_at", "2026-08-03T12:00:00Z")), subject_id=subject_id)
-    event["record_ref"] = hashed_reference("mkt-record-v1", f"record-{index}")
-    event["event_ref"] = hashed_reference("mkt-event-v1", f"event-{index}")
-    event["source"]["evidence_ref"] = hashed_reference("mkt-evidence-v1:sha256", f"evidence-{index}")
-    event["quality"]["evidence_ref"] = event["source"]["evidence_ref"]
-    event["scope"]["dimensions"]["experiment_variant"] = variant_id
-    event["measurement"].update(
-        {
-            "metric_id": metric_id,
-            "value": spec.get("value", 1),
-            "unit": spec.get("unit", "conversion"),
-            "currency": None,
-        }
-    )
-    return event
-
-
-def experiment_subject(subject_id: str) -> dict[str, object]:
-    """Build one complete subject record for experiment fixtures."""
-    return {
-        "schema_version": 1,
-        "subject_id": subject_id,
-        "kind": "lead",
-        "identity_state": "isolated",
-        "canonical_subject_id": subject_id,
-        "aliases": [subject_id],
-        "consent": [],
-        "suppression": [],
-        "identity_history": [],
-        "audience_eligible": False,
-        "eligibility_reason": "consent_unknown",
-    }
-
-
-def experiment_fixture(options: dict[str, int] | None = None) -> tuple[dict[str, object], dict[str, object], dict[str, object]]:
-    """Build balanced or deliberately adversarial experiment fixtures."""
-    fixture = load_fixture("verified-control-experiment.json")
-    settings = dict(fixture["settings"])
-    settings.update(options or {})
-    variants = (("control", settings["control_subjects"]), ("treatment", settings["treatment_subjects"]))
-    assignments: list[dict[str, object]] = []
-    events: list[dict[str, object]] = []
-    subjects: list[dict[str, object]] = []
-    index = 0
-    for variant_id, count in variants:
-        for position in range(count):
-            subject_id = hashed_reference("mkt-subj-v1", f"{variant_id}-{position}")
-            assignments.append({"unit_ref": subject_id, "variant_id": variant_id, "assigned_at": "2026-08-01T00:00:00Z"})
-            subjects.append(experiment_subject(subject_id))
-            events.append(
-                experiment_event(
-                    index,
-                    {
-                        "event_type": "impression",
-                        "metric_id": "marketing.experiment.exposures",
-                        "unit": "impression",
-                        "subject_id": subject_id,
-                        "variant_id": variant_id,
-                        "occurred_at": "2026-08-02T12:00:00Z",
-                    },
-                )
-            )
-            index += 1
-            if position < settings[f"{variant_id}_conversions"]:
-                events.append(
-                    experiment_event(
-                        index,
-                        {
-                            "event_type": "conversion",
-                            "metric_id": "marketing.conversions.total",
-                            "unit": "conversion",
-                            "subject_id": subject_id,
-                            "variant_id": variant_id,
-                        },
-                    )
-                )
-                index += 1
-            has_positive_guardrail = (
-                variant_id == "treatment" and position < settings["treatment_guardrails"]
-            )
-            if has_positive_guardrail:
-                events.append(
-                    experiment_event(
-                        index,
-                        {
-                            "event_type": "unsubscribe",
-                            "metric_id": "marketing.unsubscribes.total",
-                            "unit": "unsubscribe",
-                            "subject_id": subject_id,
-                            "variant_id": variant_id,
-                        },
-                    )
-                )
-                index += 1
-            elif position < MINIMUM_AGGREGATE_CELL_SIZE:
-                events.append(
-                    experiment_event(
-                        index,
-                        {
-                            "event_type": "unsubscribe",
-                            "metric_id": "marketing.unsubscribes.total",
-                            "unit": "unsubscribe",
-                            "value": 0,
-                            "subject_id": subject_id,
-                            "variant_id": variant_id,
-                        },
-                    )
-                )
-                index += 1
-    assignment: dict[str, object] = {
-        "schema": "aidevops.marketing-assignment-snapshot/v1",
-        "experiment_id": "landing-page-test",
-        "definition_version": 1,
-        "generated_at": "2026-08-01T00:00:00Z",
-        "assignments": assignments,
-    }
-    assignment["assignment_ref"] = typed_reference("mkt-assignment-v1:sha256", assignment)
-    definition = dict(fixture["definition"])
-    definition["assignment"]["snapshot_ref"] = assignment["assignment_ref"]
-    snapshot: dict[str, object] = {
-        "schema": "aidevops.marketing-optimization-snapshot/v1",
-        "as_of": AS_OF,
-        "events": events,
-        "subjects": subjects,
-        "sources": [
-            {
-                "source": "normalized",
-                "account_ref": ACCOUNT,
-                "status": "ready",
-                "coverage": "complete",
-                "missing_scopes": [],
-                "last_observed_at": "2026-08-03T12:00:00Z",
-                "stale_after_seconds": 30 * 86400,
-                "lag_seconds": 0,
-                "stale": False,
-                "unresolved_quarantine": 0,
-            }
-        ],
-    }
-    return definition, assignment, snapshot
-
-
-def future_cli_experiment_fixture() -> tuple[dict[str, object], dict[str, object], dict[str, object]]:
-    """Return an experiment whose trusted registration boundary is still open."""
-    definition, assignment, snapshot = experiment_fixture()
-    now = datetime.now(timezone.utc).replace(microsecond=0)
-    start = now + timedelta(days=1)
-    exposure = start + timedelta(days=1)
-    outcome = start + timedelta(days=2)
-    end = start + timedelta(days=9)
-    as_of = end + timedelta(days=5)
-
-    def stamp(value: datetime) -> str:
-        return value.isoformat().replace("+00:00", "Z")
-
-    definition["hypothesis"]["preregistered_at"] = stamp(now)
-    definition["data_policy"]["started_at"] = stamp(start)
-    definition["data_policy"]["ended_at"] = stamp(end)
-    assignment["generated_at"] = stamp(now)
-    for row in assignment["assignments"]:
-        row["assigned_at"] = stamp(start)
-    assignment.pop("assignment_ref")
-    assignment["assignment_ref"] = typed_reference("mkt-assignment-v1:sha256", assignment)
-    definition["assignment"]["snapshot_ref"] = assignment["assignment_ref"]
-    for event in snapshot["events"]:
-        occurred_at = exposure if event["measurement"]["metric_id"] == "marketing.experiment.exposures" else outcome
-        timestamp = stamp(occurred_at)
-        event["event"]["occurred_at"] = timestamp
-        for field in ("observed_at", "recorded_at", "source_observed_at", "source_recorded_at"):
-            event["source"][field] = timestamp
-    snapshot["as_of"] = stamp(as_of)
-    snapshot["sources"][0]["last_observed_at"] = stamp(outcome)
-    return definition, assignment, snapshot
-
-
-class MarketingOptimizationContractTests(unittest.TestCase):
+class MarketingOptimizationContractTests(SnapshotAttributionMixin, unittest.TestCase):
     """Exercise deterministic snapshots and immutable public-safe output."""
 
     def setUp(self) -> None:
@@ -865,367 +555,6 @@ class MarketingOptimizationContractTests(unittest.TestCase):
                         account_ref=ACCOUNT,
                         campaign_id=CAMPAIGN,
                     )
-
-    def test_live_snapshot_ignores_all_state_recorded_after_as_of(self) -> None:
-        def event(
-            source_event_id: str,
-            subject_ref: str,
-            *,
-            revision: int = 1,
-            suppressed: bool = False,
-        ) -> dict[str, object]:
-            governance: dict[str, object] = {
-                "consent": [
-                    {
-                        "purpose": "audience",
-                        "state": "granted",
-                        "lawful_basis": "synthetic-test",
-                        "effective_at": "2026-08-11T00:00:00Z",
-                    }
-                ],
-                "suppression": None,
-            }
-            if suppressed:
-                governance["suppression"] = {
-                    "state": "suppressed",
-                    "reason": "synthetic-test",
-                    "effective_at": "2026-08-11T00:00:00Z",
-                }
-            return {
-                "source_event_id": source_event_id,
-                "revision": revision,
-                "event_type": "conversion",
-                "occurred_at": "2026-08-11T00:00:00Z",
-                "correction_of": None,
-                "subject": {
-                    "kind": "lead",
-                    "identity_state": "isolated",
-                    "source_ref": subject_ref,
-                    "candidate_refs": [],
-                },
-                "scope": {"campaign_id": CAMPAIGN, "channel": "search"},
-                "measurement": {
-                    "metric_id": "marketing.conversions.total",
-                    "value": revision,
-                    "unit": "conversion",
-                    "aggregation": "sum",
-                    "currency": None,
-                },
-                "quality": {
-                    "confidence": "high",
-                    "completeness": "complete",
-                    "source_type": "fixture",
-                    "collected_by": "synthetic-fixture",
-                },
-                "governance": governance,
-            }
-
-        def adapter_result(
-            events: list[dict[str, object]],
-            cursor: str,
-            observed_at: str,
-            *,
-            errors: list[dict[str, object]] | None = None,
-        ) -> AdapterResult:
-            batch = {
-                "source": "normalized",
-                "account_ref": ACCOUNT,
-                "cursor": cursor,
-                "observed_at": observed_at,
-                "coverage": "complete",
-                "missing_scopes": [],
-                "events": events,
-            }
-            return AdapterResult(
-                batch=batch,
-                errors=errors or [],
-                raw_bytes=json.dumps(batch, sort_keys=True).encode("utf-8"),
-                suffix=".json",
-            )
-
-        initial_epoch = int(
-            datetime.fromisoformat("2026-08-11T01:00:00+00:00").timestamp()
-        )
-        future_epoch = int(
-            datetime.fromisoformat("2026-08-12T01:00:00+00:00").timestamp()
-        )
-        with MarketingPerformanceStore.open(self.repo, provision=True) as store:
-            with mock.patch(
-                "_performance_store_leases.time.time",
-                return_value=initial_epoch,
-            ):
-                with mock.patch(
-                    "_performance_store_ingest.time.time",
-                    return_value=initial_epoch,
-                ):
-                    with mock.patch(
-                        "_performance_store_ingest.utc_now",
-                        return_value="2026-08-11T01:00:00Z",
-                    ):
-                        store.ingest(
-                            "normalized",
-                            adapter_result(
-                                [
-                                    event("event-a", "subject-a"),
-                                    event("event-b", "subject-b"),
-                                ],
-                                "initial-cursor",
-                                "2026-08-11T00:30:00Z",
-                                errors=[
-                                    {
-                                        "index": "synthetic",
-                                        "reason": "synthetic quarantine",
-                                        "source_event_id": "quarantined-event",
-                                    }
-                                ],
-                            ),
-                        )
-            subjects = [
-                str(row["subject_id"])
-                for row in store.connection.execute(
-                    "SELECT DISTINCT subject_id FROM events ORDER BY subject_id"
-                )
-            ]
-            quarantine_ref = str(
-                store.connection.execute(
-                    "SELECT quarantine_ref FROM quarantine"
-                ).fetchone()[0]
-            )
-
-        boundary = "2026-08-11T12:00:00Z"
-        before = snapshot_from_repo(
-            self.repo,
-            as_of=boundary,
-            account_ref=ACCOUNT,
-            campaign_id=CAMPAIGN,
-        )
-
-        with MarketingPerformanceStore.open(self.repo) as store:
-            with mock.patch(
-                "_performance_store_leases.time.time",
-                return_value=future_epoch,
-            ):
-                with mock.patch(
-                    "_performance_store_ingest.time.time",
-                    return_value=future_epoch,
-                ):
-                    with mock.patch(
-                        "_performance_store_ingest.utc_now",
-                        return_value="2026-08-12T01:00:00Z",
-                    ):
-                        store.ingest(
-                            "normalized",
-                            adapter_result(
-                                [
-                                    event(
-                                        "event-a",
-                                        "subject-a",
-                                        revision=2,
-                                        suppressed=True,
-                                    )
-                                ],
-                                "future-cursor",
-                                "2026-08-11T00:45:00Z",
-                            ),
-                        )
-            with mock.patch(
-                "_performance_reporting_reconciliation.utc_now",
-                return_value="2026-08-12T01:00:00Z",
-            ):
-                with mock.patch(
-                    "_performance_reporting_queries.utc_now",
-                    return_value="2026-08-12T01:00:00Z",
-                ):
-                    PerformanceReporting(store).reconcile(
-                        {
-                            "schema": "aidevops.marketing-performance-reconciliation/v1",
-                            "actions": [
-                                {
-                                    "action": "link",
-                                    "canonical_subject_id": subjects[0],
-                                    "member_subject_id": subjects[1],
-                                    "effective_at": "2026-08-11T00:00:00Z",
-                                    "evidence_ref": "synthetic-owner-link",
-                                },
-                                {
-                                    "action": "resolve_quarantine",
-                                    "quarantine_ref": quarantine_ref,
-                                    "resolution": "discarded",
-                                    "effective_at": "2026-08-11T00:00:00Z",
-                                    "evidence_ref": "synthetic-owner-resolution",
-                                },
-                            ],
-                        }
-                    )
-
-        after = snapshot_from_repo(
-            self.repo,
-            as_of=boundary,
-            account_ref=ACCOUNT,
-            campaign_id=CAMPAIGN,
-        )
-
-        self.assertEqual(before, after)
-        self.assertEqual(2, len(after.events))
-        self.assertEqual(2, len(after.subjects))
-        self.assertEqual("partial", after.sources[0]["status"])
-        self.assertEqual(1, after.sources[0]["unresolved_quarantine"])
-
-    def test_live_snapshot_preserves_corrected_outcome_refund_chronology(self) -> None:
-        def raw_event(
-            source_event_id: str,
-            event_type: str,
-            occurred_at: str,
-            subject_ref: str,
-            outcome_id: str,
-            metric_id: str,
-            value: int,
-            *,
-            correction_of: str | None = None,
-        ) -> dict[str, object]:
-            return {
-                "source_event_id": source_event_id,
-                "revision": 1,
-                "event_type": event_type,
-                "occurred_at": occurred_at,
-                "correction_of": correction_of,
-                "subject": {
-                    "kind": "lead",
-                    "identity_state": "isolated",
-                    "source_ref": subject_ref,
-                    "candidate_refs": [],
-                },
-                "scope": {
-                    "campaign_id": CAMPAIGN,
-                    "channel": "search",
-                    "outcome_id": outcome_id,
-                },
-                "measurement": {
-                    "metric_id": metric_id,
-                    "value": value,
-                    "unit": "currency",
-                    "aggregation": "sum",
-                    "currency": "USD",
-                },
-                "quality": {
-                    "confidence": "high",
-                    "completeness": "complete",
-                    "source_type": "fixture",
-                    "collected_by": "synthetic-fixture",
-                },
-                "governance": {"consent": [], "suppression": None},
-            }
-
-        events: list[dict[str, object]] = []
-        for index in range(MINIMUM_AGGREGATE_CELL_SIZE):
-            target_ref = f"revenue-{index}"
-            subject_ref = f"subject-{index}"
-            outcome_id = f"order-{index}"
-            events.extend(
-                (
-                    raw_event(
-                        target_ref,
-                        "revenue",
-                        "2026-08-01T12:00:00Z",
-                        subject_ref,
-                        outcome_id,
-                        "marketing.revenue.gross",
-                        100,
-                    ),
-                    raw_event(
-                        f"correction-{index}",
-                        "correction",
-                        "2026-08-03T12:00:00Z",
-                        subject_ref,
-                        outcome_id,
-                        "marketing.revenue.gross",
-                        120,
-                        correction_of=target_ref,
-                    ),
-                    raw_event(
-                        f"refund-{index}",
-                        "refund",
-                        "2026-08-02T12:00:00Z",
-                        subject_ref,
-                        outcome_id,
-                        "marketing.revenue.refunded",
-                        25,
-                    ),
-                )
-            )
-        batch = {
-            "source": "normalized",
-            "account_ref": ACCOUNT,
-            "cursor": "corrected-refund-cursor",
-            "observed_at": "2026-08-04T00:00:00Z",
-            "coverage": "complete",
-            "missing_scopes": [],
-            "events": events,
-        }
-        result = AdapterResult(
-            batch=batch,
-            errors=[],
-            raw_bytes=json.dumps(batch, sort_keys=True).encode("utf-8"),
-            suffix=".json",
-        )
-        recorded_epoch = int(datetime.fromisoformat("2026-08-04T01:00:00+00:00").timestamp())
-        with MarketingPerformanceStore.open(self.repo, provision=True) as store:
-            with mock.patch("_performance_store_leases.time.time", return_value=recorded_epoch):
-                with mock.patch("_performance_store_ingest.time.time", return_value=recorded_epoch):
-                    with mock.patch(
-                        "_performance_store_ingest.utc_now",
-                        return_value="2026-08-04T01:00:00Z",
-                    ):
-                        store.ingest("normalized", result)
-
-        snapshot = snapshot_from_repo(
-            self.repo,
-            as_of=AS_OF,
-            account_ref=ACCOUNT,
-            campaign_id=CAMPAIGN,
-        )
-        projection = build_attribution(
-            snapshot,
-            attribution_request(
-                outcome_metric_id="marketing.revenue.gross",
-                account_ref=ACCOUNT,
-                campaign_id=CAMPAIGN,
-            ),
-        )
-
-        self.assertEqual(120 * MINIMUM_AGGREGATE_CELL_SIZE, projection["outcomes"]["gross_value"])
-        self.assertEqual(25 * MINIMUM_AGGREGATE_CELL_SIZE, projection["outcomes"]["refund_value"])
-        self.assertEqual(95 * MINIMUM_AGGREGATE_CELL_SIZE, projection["outcomes"]["net_value"])
-        self.assertEqual(0, projection["coverage"]["unmatched_refunds"])
-
-    def test_last_touch_attribution_is_deterministic_and_schema_valid(self) -> None:
-        snapshot = snapshot_from_document(aggregate_analysis_document())
-
-        first = build_attribution(snapshot, attribution_request())
-        second = build_attribution(snapshot, attribution_request())
-
-        self.assertEqual(first, second)
-        self.assertEqual(MINIMUM_AGGREGATE_CELL_SIZE, first["outcomes"]["attributed_count"])
-        self.assertTrue(all(item["suppressed"] for item in first["allocations"]))
-        self.assertTrue(all(item["touchpoint_ref"] is None for item in first["allocations"]))
-        self.assertEqual("complete", first["run"]["status"])
-        self.assertEqual("observational_only", first["causal_assessment"]["status"])
-        self.assert_schema(ATTRIBUTION_SCHEMA, first)
-
-    def test_small_attribution_cells_hide_counts_values_and_dimensions(self) -> None:
-        snapshot = snapshot_from_document(analysis_document())
-
-        projection = build_attribution(snapshot, attribution_request(minimum_cell_size=10))
-
-        self.assertTrue(projection["outcomes"]["suppressed"])
-        self.assertIsNone(projection["outcomes"]["eligible_count"])
-        self.assertIsNone(projection["allocations"][0]["channel"])
-        self.assertIsNone(projection["allocations"][0]["value"])
-        self.assertIsNone(projection["coverage"]["suppressed_allocations"])
-        self.assertIsNone(projection["coverage"]["late_events"])
-        self.assertIsNone(projection["coverage"]["unmatched_refunds"])
-        self.assertEqual("low", projection["uncertainty"]["data_confidence"])
-        self.assert_schema(ATTRIBUTION_SCHEMA, projection)
 
     def test_external_attribution_validation_enforces_privacy_invariants(self) -> None:
         suppressed = build_attribution(
